Update Azure resource and service interfaces
