+++ azure-toolkit-libs/azure-toolkit-common-lib/src/main/java/com/microsoft/azure/toolkit/lib/SubscriptionScoped.java
@@ -22,7 +22,7 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
-public abstract class SubscriptionScoped<T extends AzureService> {
+public abstract class SubscriptionScoped<T> {
     @Nonnull
     private final Function<List<Subscription>, T> creator;
     @Nullable
